Use the same syncing call everywhere we can
This also means we capture details for unconfigured launches, which
previously we did not.
